It now works both with and without javascript.
The errors are now defined in one place, and called with javascript.

You can change it back, but I really think the banner on top was a bad idea.  The colors didn't match the theme and, more importantly I didn't see it at first and was wondering why the page wasn't responding.